move tx tests to new file; remove unnecessary pub on test modules
