+++ mm2src/coins/sia/src/tests/mod.rs
@@ -1,3 +1,4 @@
-pub mod encoding;
-pub mod serde;
-pub mod spend_policy;
+mod encoding;
+mod serde;
+mod spend_policy;
+mod transaction;
